+++ src/mako/database/query/compilers/MySQL.php
@@ -22,6 +22,7 @@ class MySQL extends Compiler
 	/**
 	 * {@inheritDoc}
 	 */
+	#[Override]
 	protected static string $dateFormat = 'Y-m-d H:i:s';
 
 	/**