+++ src/mako/database/query/compilers/SQLite.php
@@ -20,6 +20,7 @@ class SQLite extends Compiler
 	/**
 	 * {@inheritDoc}
 	 */
+	#[Override]
 	protected static string $dateFormat = 'Y-m-d H:i:s';
 
 	/**